refs #24 Divide class for hatena api client
