+++ user/views.py
@@ -1,9 +1,8 @@
 import json
-import requests
-from requests_oauthlib import OAuth1
 from django.http import HttpResponse, JsonResponse
 from allauth.socialaccount.models import SocialAccount, SocialToken
 from hatenaprovider.provider import HatenaProvider
+from modules.hatena.client import HatenaAPIClient
 
 
 def my(request):
@@ -29,36 +28,8 @@ def bookmark(request):
 
     token = SocialToken.objects.all().select_related('account', 'app').filter(app__provider=HatenaProvider.id).filter(account__user_id=request.user.id)[0]
 
-    api = HatenaAPI(token.app.client_id, token.app.secret, token.token, token.token_secret)
+    api = HatenaAPIClient(token.app.client_id, token.app.secret, token.token, token.token_secret)
 
     params = json.loads(request.body)
     res = api.bookmark(params['url'], params['comment'])
     return JsonResponse(res)
-
-
-class HatenaAPI():
-    base_url = 'http://api.b.hatena.ne.jp/1'
-
-    def __init__(self, consumer_key, consumer_secret, oauth_token, oauth_token_secret):
-        self.oauth = OAuth1(
-            consumer_key,
-            client_secret=consumer_secret,
-            resource_owner_key=oauth_token,
-            resource_owner_secret=oauth_token_secret)
-
-    def bookmark(self, target_url, comment):
-        path = '/my/bookmark'
-        url = self.base_url + path
-        params = {
-            'url': target_url,
-            'comment': comment,
-        }
-
-        response = getattr(requests, 'post')(url,
-                                             auth=self.oauth,
-                                             headers=dict(),
-                                             params=params)
-#        import pdb; pdb.set_trace()
-        response.raise_for_status()
-
-        return response.json()
